+++ Sample/Sample/Menu Controllers/KittyViewController.swift
@@ -19,13 +19,16 @@
 import UIKit
 import InteractiveSideMenu
 
-/*
+/**
  KittyViewController is a controller relevant one of the side menu items. To indicate this it adopts `SideMenuItemContent` protocol.
  */
-class KittyViewController: UIViewController, SideMenuItemContent {
+class KittyViewController: UIViewController, SideMenuItemContent, Storyboardable {
 
-    /* Show side menu on menu button click
-     */
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return .lightContent
+    }
+
+    // Show side menu on menu button click
     @IBAction func openMenu(_ sender: UIButton) {
         showSideMenu()
     }