+++ Sample/Sample/Menu Controllers/TabBarViewController.swift
@@ -19,42 +19,27 @@
 import UIKit
 import InteractiveSideMenu
 
-/*
+/**
  TabBarViewController is a controller relevant one of the side menu items. To indicate this it adopts `SideMenuItemContent` protocol.
  */
-class TabBarViewController: UITabBarController, SideMenuItemContent {
+class TabBarViewController: UITabBarController, SideMenuItemContent, Storyboardable {
+
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return .lightContent
+    }
+
+    @IBAction func openMenu(_ sender: UIBarButtonItem) {
+        showSideMenu()
+    }
 
 }
 
-/*
+/**
  The first controller of tab bar.
  */
-class FirstViewController: UIViewController {
+class FirstViewController: UIViewController { }
 
-    /*
-     Show menu on click if connected tab bar controller adopts proper protocol.
-     */
-    @IBAction func openMenu(_ sender: UIButton) {
-
-        if let menuItemViewController = self.tabBarController as? SideMenuItemContent {
-            menuItemViewController.showSideMenu()
-        }
-    }
-}
-
-/*
+/**
  The second controller of tab bar.
  */
-class SecondViewController: UIViewController {
-
-    /*
-     Show menu on click if connected tab bar controller adopts proper protocol.
-     */
-    @IBAction func openMenu(_ sender: UIButton) {
-
-        if let menuItemViewController = self.tabBarController as? SideMenuItemContent {
-            menuItemViewController.showSideMenu()
-        }
-    }
-
-}
+class SecondViewController: UIViewController { }