+++ Sources/MenuContainerViewController.swift
@@ -23,8 +23,23 @@ import UIKit
  */
 open class MenuContainerViewController: UIViewController {
 
-    // Describes if side menu is shown or not.
-    var isShown = false
+    open override var preferredStatusBarStyle: UIStatusBarStyle {
+        return currentContentViewController?.preferredStatusBarStyle ?? .lightContent
+    }
+
+    fileprivate weak var currentContentViewController: UIViewController?
+    fileprivate var navigationMenuTransitionDelegate: MenuTransitioningDelegate!
+
+    /**
+     Flag indicating if the side menu is being shown.
+     */
+    fileprivate var isShown = false
+
+    public var shadowOptions = SideMenuItemShadow() {
+        didSet {
+            navigationMenuTransitionDelegate?.shadowOptions = shadowOptions
+        }
+    }
 
     /**
      The view controller for side menu.
@@ -58,51 +73,23 @@ open class MenuContainerViewController: UIViewController {
      */
     public var contentViewControllers = [UIViewController]()
 
-    /**
-     Shows left side menu.
-     */
-    public func showSideMenu() {
-        presentNavigationMenu()
-    }
-
-    /**
-     Hides left side menu.
-     Controller from the right side will be visible.
-     */
-    public func hideSideMenu() {
-        dismissNavigationMenu()
-    }
-
-    /**
-     Embeds menu item content view controller.
-
-     - parameter selectedContentVC: The view controller to be embedded.
-     */
-    public func selectContentViewController(_ selectedContentVC: UIViewController) {
-        if let currentContentVC = currentContentViewController {
-            if currentContentVC != selectedContentVC {
-                currentContentVC.view.removeFromSuperview()
-                currentContentVC.removeFromParentViewController()
-                setCurrentView(selectedContentVC)
-            }
-        } else {
-            setCurrentView(selectedContentVC)
-        }
-    }
 
     // MARK: - Controller lifecycle
     //
     override open func viewDidLoad() {
         super.viewDidLoad()
 
-        navigationMenuTransitionDelegate = MenuTransitioningDelegate(interactiveTransition: MenuInteractiveTransition(
+        let interactiveTransition = MenuInteractiveTransition(
+            shadowOptions: shadowOptions,
             presentAction: { [unowned self] in
                 self.presentNavigationMenu()
             },
             dismissAction: { [unowned self] in
                 self.dismissNavigationMenu()
             }
-        ))
+        )
+
+        navigationMenuTransitionDelegate = MenuTransitioningDelegate(interactiveTransition: interactiveTransition)
 
         let screenEdgePanRecognizer = UIScreenEdgePanGestureRecognizer(
             target: navigationMenuTransitionDelegate.interactiveTransition,
@@ -131,18 +118,53 @@ open class MenuContainerViewController: UIViewController {
             }
         }, completion: nil)
     }
+}
 
-    // MARK: - Private
-    //
-    private weak var currentContentViewController: UIViewController?
-    private var navigationMenuTransitionDelegate: MenuTransitioningDelegate!
+// MARK: - Public
+extension MenuContainerViewController {
 
+    /**
+     Shows left side menu.
+     */
+    public func showSideMenu() {
+        presentNavigationMenu()
+    }
+
+    /**
+     Hides left side menu.
+     Controller from the right side will be visible.
+     */
+    public func hideSideMenu() {
+        dismissNavigationMenu()
+    }
+
+    /**
+     Embeds menu item content view controller.
+
+     - parameter selectedContentVC: The view controller to be embedded.
+     */
+    public func selectContentViewController(_ selectedContentVC: UIViewController) {
+        if let currentContentVC = currentContentViewController {
+            if currentContentVC != selectedContentVC {
+                currentContentVC.view.removeFromSuperview()
+                currentContentVC.removeFromParentViewController()
+
+                setCurrentView(selectedContentVC)
+            }
+        } else {
+            setCurrentView(selectedContentVC)
+        }
+    }
+}
+
+// MARK: - Private
+fileprivate extension MenuContainerViewController {
     /**
      Adds proper content view controller as a child.
 
      - parameter selectedContentVC: The view controller to be added.
      */
-    private func setCurrentView(_ selectedContentVC: UIViewController) {
+    func setCurrentView(_ selectedContentVC: UIViewController) {
         addChildViewController(selectedContentVC)
         view.addSubviewWithFullSizeConstraints(view: selectedContentVC.view)
         currentContentViewController = selectedContentVC
@@ -151,7 +173,7 @@ open class MenuContainerViewController: UIViewController {
     /**
      Presents left side menu.
      */
-    private func presentNavigationMenu() {
+    func presentNavigationMenu() {
         if menuViewController == nil {
             fatalError("Invalid `menuViewController` value. It should not be nil")
         }
@@ -162,7 +184,7 @@ open class MenuContainerViewController: UIViewController {
     /**
      Dismisses left side menu.
      */
-    private func dismissNavigationMenu() {
+    func dismissNavigationMenu() {
         self.dismiss(animated: true, completion: nil)
         isShown = false
     }
@@ -177,7 +199,10 @@ extension UIView {
         view.translatesAutoresizingMaskIntoConstraints = false
         insertSubview(view, at: atIndex)
 
-        addConstraints(NSLayoutConstraint.constraints(withVisualFormat: "V:|[view]|", options: NSLayoutFormatOptions(), metrics: nil, views: ["view": view]))
-        addConstraints(NSLayoutConstraint.constraints(withVisualFormat: "H:|[view]|", options: NSLayoutFormatOptions(), metrics: nil, views: ["view": view]))
+        let top = view.topAnchor.constraint(equalTo: self.topAnchor)
+        let leading = view.leadingAnchor.constraint(equalTo: self.leadingAnchor)
+        let trailing = self.trailingAnchor.constraint(equalTo: view.trailingAnchor)
+        let bottom = self.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+        NSLayoutConstraint.activate([top, leading, trailing, bottom])
     }
 }